Proper fix for spacing in concatenated strings.
