--- intelij-plugin/src/main/gen/tech/pantheon/yanginator/plugin/psi/YangPublicSep.java
@@ -1,16 +1,0 @@
-// This is a generated file. Not intended for manual editing.
-package tech.pantheon.yanginator.plugin.psi;
-
-import org.jetbrains.annotations.NotNull;
-
-import java.util.List;
-
-public interface YangPublicSep extends YangNamedElement {
-
-  @NotNull
-  List<YangLineComment> getLineCommentList();
-
-  @NotNull
-  List<YangComment> getCommentList();
-
-}
