--- intelij-plugin/src/main/gen/tech/pantheon/yanginator/plugin/psi/impl/YangPublicSepImpl.java
@@ -1,43 +1,0 @@
-// This is a generated file. Not intended for manual editing.
-package tech.pantheon.yanginator.plugin.psi.impl;
-
-import com.intellij.lang.ASTNode;
-import com.intellij.psi.PsiElementVisitor;
-import com.intellij.psi.util.PsiTreeUtil;
-import org.jetbrains.annotations.NotNull;
-import tech.pantheon.yanginator.plugin.psi.YangComment;
-import tech.pantheon.yanginator.plugin.psi.YangLineComment;
-import tech.pantheon.yanginator.plugin.psi.YangPublicSep;
-import tech.pantheon.yanginator.plugin.psi.YangVisitor;
-
-import java.util.List;
-
-public class YangPublicSepImpl extends YangNamedElementImpl implements YangPublicSep {
-
-  public YangPublicSepImpl(@NotNull ASTNode node) {
-    super(node);
-  }
-
-  public void accept(@NotNull YangVisitor visitor) {
-    visitor.visitPublicSep(this);
-  }
-
-  @Override
-  public void accept(@NotNull PsiElementVisitor visitor) {
-    if (visitor instanceof YangVisitor) accept((YangVisitor)visitor);
-    else super.accept(visitor);
-  }
-
-  @Override
-  @NotNull
-  public List<YangLineComment> getLineCommentList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, YangLineComment.class);
-  }
-
-  @Override
-  @NotNull
-  public List<YangComment> getCommentList() {
-    return PsiTreeUtil.getChildrenOfTypeAsList(this, YangComment.class);
-  }
-
-}
